+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [0.1.5] - 2024-01-30
+- Added callbacks to cache:
+    - Create/POST, refresh cache after successful POST request
+    - Update/PUT, refresh cache after successful PUT request
+    - Destroy/DELETE, invalidate cache after successful DELETE request
+- Fixed issue with generator for SQLCache strategy tables
 ## [0.1.4] - 2024-12-20
 - CI Improvements
     - Added annotations